additional scripting

+++ scripts/clean.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
-COMMAND="rm -rf ${DIRECTORY}/target/*";
-echo ${COMMAND};
-${COMMAND}
+if [ -e ${DIRECTORY}/target ]; then
+	COMMAND="rm -rf ${DIRECTORY}/target/*";
+	echo ${COMMAND};
+	${COMMAND}
+fi
